[N06] Not declaring uint as uint256 (fix) (#593)
* Explicitly declare all instances of uint as uint256

+++ eth-contracts/contracts/test/MockDelegateManager.sol
@@ -20,8 +20,8 @@ contract MockDelegateManager is InitializableV2 {
     // Test only function
     function testProcessClaim(
         address _claimer,
-        uint _totalLockedForSP
-    ) external returns (uint) {
+        uint256 _totalLockedForSP
+    ) external returns (uint256) {
         _requireIsInitialized();
 
         ClaimsManager claimsManager = ClaimsManager(
